# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ScreenGenie.ai API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardStats {
  totalJobs: number;
  activeJobs: number;
  totalCandidates: number;
  pendingInterviews: number;
  completedInterviews: number;
  hiredCandidates: number;
}

export interface User {
  id: number;
  email: string;
  name: string;
  company: string;
  createdAt?: string;
}

export interface CreateUser {
  email: string;
  name: string;
  company: string;
}

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus];

export const JobStatus = {
  OPEN: "OPEN",
  CLOSED: "CLOSED",
  DRAFT: "DRAFT",
} as const;

export interface Job {
  id: number;
  hrId: number;
  title: string;
  description: string;
  skills: string;
  status: JobStatus;
  createdAt?: string;
}

export type CreateJobStatus =
  (typeof CreateJobStatus)[keyof typeof CreateJobStatus];

export const CreateJobStatus = {
  OPEN: "OPEN",
  CLOSED: "CLOSED",
  DRAFT: "DRAFT",
} as const;

export interface CreateJob {
  hrId: number;
  title: string;
  description: string;
  skills: string;
  status?: CreateJobStatus;
}

export type CandidateStatus =
  (typeof CandidateStatus)[keyof typeof CandidateStatus];

export const CandidateStatus = {
  PENDING: "PENDING",
  INVITED: "INVITED",
  SCHEDULED: "SCHEDULED",
  INTERVIEWED: "INTERVIEWED",
  HIRED: "HIRED",
  REJECTED: "REJECTED",
} as const;

export interface Candidate {
  id: number;
  jobId: number;
  name: string;
  email: string;
  phone?: string;
  status: CandidateStatus;
  score?: number | null;
  resumeUrl?: string | null;
  createdAt?: string;
}

export type CreateCandidateStatus =
  (typeof CreateCandidateStatus)[keyof typeof CreateCandidateStatus];

export const CreateCandidateStatus = {
  PENDING: "PENDING",
  INVITED: "INVITED",
  SCHEDULED: "SCHEDULED",
  INTERVIEWED: "INTERVIEWED",
  HIRED: "HIRED",
  REJECTED: "REJECTED",
} as const;

export interface CreateCandidate {
  jobId: number;
  name: string;
  email: string;
  phone?: string;
  status?: CreateCandidateStatus;
  score?: number | null;
  resumeUrl?: string | null;
}

export type UpdateCandidateStatus =
  (typeof UpdateCandidateStatus)[keyof typeof UpdateCandidateStatus];

export const UpdateCandidateStatus = {
  PENDING: "PENDING",
  INVITED: "INVITED",
  SCHEDULED: "SCHEDULED",
  INTERVIEWED: "INTERVIEWED",
  HIRED: "HIRED",
  REJECTED: "REJECTED",
} as const;

export interface UpdateCandidate {
  name?: string;
  email?: string;
  phone?: string;
  status?: UpdateCandidateStatus;
  score?: number | null;
  resumeUrl?: string | null;
}

export type InterviewStatus =
  (typeof InterviewStatus)[keyof typeof InterviewStatus];

export const InterviewStatus = {
  SCHEDULED: "SCHEDULED",
  IN_PROGRESS: "IN_PROGRESS",
  COMPLETED: "COMPLETED",
  CANCELLED: "CANCELLED",
} as const;

export interface Interview {
  id: number;
  candidateId: number;
  scheduledAt: string;
  status: InterviewStatus;
  attempts: number;
  transcript?: string | null;
  feedback?: string | null;
  createdAt?: string;
}

export type CreateInterviewStatus =
  (typeof CreateInterviewStatus)[keyof typeof CreateInterviewStatus];

export const CreateInterviewStatus = {
  SCHEDULED: "SCHEDULED",
  IN_PROGRESS: "IN_PROGRESS",
  COMPLETED: "COMPLETED",
  CANCELLED: "CANCELLED",
} as const;

export interface CreateInterview {
  candidateId: number;
  scheduledAt: string;
  status?: CreateInterviewStatus;
}

export type UpdateInterviewStatus =
  (typeof UpdateInterviewStatus)[keyof typeof UpdateInterviewStatus];

export const UpdateInterviewStatus = {
  SCHEDULED: "SCHEDULED",
  IN_PROGRESS: "IN_PROGRESS",
  COMPLETED: "COMPLETED",
  CANCELLED: "CANCELLED",
} as const;

export interface UpdateInterview {
  scheduledAt?: string;
  status?: UpdateInterviewStatus;
  attempts?: number;
  transcript?: string | null;
  feedback?: string | null;
}

export type GetCandidatesParams = {
  jobId?: number;
};

export type GetInterviewsParams = {
  candidateId?: number;
};