# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ScreenGenie.ai API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: users
    description: HR User operations
  - name: jobs
    description: Job operations
  - name: candidates
    description: Candidate operations
  - name: interviews
    description: Interview operations
  - name: dashboard
    description: Dashboard statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: List all HR users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new HR user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUser"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUser"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /jobs:
    get:
      operationId: getJobs
      tags: [jobs]
      summary: List all jobs
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a new job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJob"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    put:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJob"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /candidates:
    get:
      operationId: getCandidates
      tags: [candidates]
      summary: List all candidates
      parameters:
        - name: jobId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Create a new candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCandidate"
      responses:
        "201":
          description: Created candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get a candidate by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    put:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCandidate"
      responses:
        "200":
          description: Updated candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /interviews:
    get:
      operationId: getInterviews
      tags: [interviews]
      summary: List all interviews
      parameters:
        - name: candidateId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of interviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Interview"
    post:
      operationId: createInterview
      tags: [interviews]
      summary: Create a new interview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInterview"
      responses:
        "201":
          description: Created interview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
  /interviews/{id}:
    get:
      operationId: getInterview
      tags: [interviews]
      summary: Get an interview by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Interview details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
    put:
      operationId: updateInterview
      tags: [interviews]
      summary: Update an interview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInterview"
      responses:
        "200":
          description: Updated interview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    DashboardStats:
      type: object
      properties:
        totalJobs:
          type: integer
        activeJobs:
          type: integer
        totalCandidates:
          type: integer
        pendingInterviews:
          type: integer
        completedInterviews:
          type: integer
        hiredCandidates:
          type: integer
      required:
        - totalJobs
        - activeJobs
        - totalCandidates
        - pendingInterviews
        - completedInterviews
        - hiredCandidates
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        company:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - company
    CreateUser:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        company:
          type: string
      required:
        - email
        - name
        - company
    Job:
      type: object
      properties:
        id:
          type: integer
        hrId:
          type: integer
        title:
          type: string
        description:
          type: string
        skills:
          type: string
        status:
          type: string
          enum: [OPEN, CLOSED, DRAFT]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - hrId
        - title
        - description
        - skills
        - status
    CreateJob:
      type: object
      properties:
        hrId:
          type: integer
        title:
          type: string
        description:
          type: string
        skills:
          type: string
        status:
          type: string
          enum: [OPEN, CLOSED, DRAFT]
      required:
        - hrId
        - title
        - description
        - skills
    Candidate:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [PENDING, INVITED, SCHEDULED, INTERVIEWED, HIRED, REJECTED]
        score:
          type: integer
          nullable: true
        resumeUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - jobId
        - name
        - email
        - status
    CreateCandidate:
      type: object
      properties:
        jobId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [PENDING, INVITED, SCHEDULED, INTERVIEWED, HIRED, REJECTED]
        score:
          type: integer
          nullable: true
        resumeUrl:
          type: string
          nullable: true
      required:
        - jobId
        - name
        - email
    UpdateCandidate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [PENDING, INVITED, SCHEDULED, INTERVIEWED, HIRED, REJECTED]
        score:
          type: integer
          nullable: true
        resumeUrl:
          type: string
          nullable: true
    Interview:
      type: object
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [SCHEDULED, IN_PROGRESS, COMPLETED, CANCELLED]
        attempts:
          type: integer
        transcript:
          type: string
          nullable: true
        feedback:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - candidateId
        - scheduledAt
        - status
        - attempts
    CreateInterview:
      type: object
      properties:
        candidateId:
          type: integer
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [SCHEDULED, IN_PROGRESS, COMPLETED, CANCELLED]
      required:
        - candidateId
        - scheduledAt
    UpdateInterview:
      type: object
      properties:
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [SCHEDULED, IN_PROGRESS, COMPLETED, CANCELLED]
        attempts:
          type: integer
        transcript:
          type: string
          nullable: true
        feedback:
          type: string
          nullable: true