# lib/api-zod/src/generated/types/candidate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ScreenGenie.ai API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CandidateStatus } from "./candidateStatus";

export interface Candidate {
  id: number;
  jobId: number;
  name: string;
  email: string;
  phone?: string;
  status: CandidateStatus;
  score?: number | null;
  resumeUrl?: string | null;
  createdAt?: Date;
}