# lib/api-zod/src/generated/types/createCandidate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ScreenGenie.ai API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCandidateStatus } from "./createCandidateStatus";

export interface CreateCandidate {
  jobId: number;
  name: string;
  email: string;
  phone?: string;
  status?: CreateCandidateStatus;
  score?: number | null;
  resumeUrl?: string | null;
}