# lib/api-zod/src/generated/types/createJob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ScreenGenie.ai API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateJobStatus } from "./createJobStatus";

export interface CreateJob {
  hrId: number;
  title: string;
  description: string;
  skills: string;
  status?: CreateJobStatus;
}