# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ScreenGenie.ai API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalJobs: number;
  activeJobs: number;
  totalCandidates: number;
  pendingInterviews: number;
  completedInterviews: number;
  hiredCandidates: number;
}