# lib/api-zod/src/generated/types/updateInterview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ScreenGenie.ai API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateInterviewStatus } from "./updateInterviewStatus";

export interface UpdateInterview {
  scheduledAt?: Date;
  status?: UpdateInterviewStatus;
  attempts?: number;
  transcript?: string | null;
  feedback?: string | null;
}